microservice_profession_user/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
python-dotenv
Flask-DotEnv
alembic
autoenv
blinker
certifi
cffi
chardet
click
cryptography
dnspython
dominate
Flask-Login
Flask-Uploads
Flask-WTF
idna
itsdangerous
Jinja2
Mako
MarkupSafe
marshmallow
passlib
protobuf
pycparser
PyMySQL
python-dateutil
requests
python-editor
six
SQLAlchemy
urllib3
visitor
Werkzeug
WTForms
